Fix script

+++ scripts/copy_cache_to_build.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 
-for p in .package-build-cache/*/build; do
-    mkdir -p packages/${p:21}
-    rm -rf packages/${p:21}/*
-    cp -a $p/* packages/${p:21}
-done
+if [ -d .package-build-cache ]
+then
+    for p in .package-build-cache/*/build; do
+        mkdir -p packages/${p:21}
+        rm -rf packages/${p:21}/*
+        cp -a $p/* packages/${p:21}
+    done
+fi
